Create tests (#1)
* Implement first element test methods

* add testing dependencies

* add first unit test file (tbc)

* attempting first unit tests + CI

* Run unit tests on PR's

* add pyComposite as dependency

* black/flake8 fixes

* fix coverage

* add codecov action

* add `_getRandomNodeCoords` methods for both quad elements

* add .coveragerc

* add shape function derivative tests

* Add shape function sum test

* remove extraneous code

* Add README badges

* switch to -e pip install to fix codecov

+++ FEMpy/LagrangePoly.py
@@ -100,7 +100,7 @@ def LagrangePoly2d(x, y, n):
     yp = y.flatten()
     Nx = LagrangePoly1d(xp, n)
     Ny = LagrangePoly1d(yp, n)
-    N = np.zeros((len(xp), n ** 2))
+    N = np.zeros((len(xp), n ** 2), dtype=xp.dtype)
     for i in range(len(xp)):
         for j in range(n):
             for k in range(n):
